# kernel-openapi.yaml
openapi: 3.1.0
info:
  title: Institutional Kernel API
  version: "1.1.0"
  description: >
    API HTTP del Institutional Kernel de Mi Rotaract. Fuente única de verdad
    para identidad, organizaciones jerárquicas, membresías, períodos, cargos
    y autoridades, autorización contextual, solicitudes de ingreso,
    transferencias entre organizaciones y registro de módulos externos.
    Este documento es un contrato de datos derivado de `kernel-spec.md`
    (secciones 4, 5, 9 y 10). No reemplaza la especificación; ante cualquier
    conflicto, `kernel-spec.md` es la fuente normativa. El proceso de
    compatibilidad y publicación está definido en
    `kernel-openapi-contract-plan.md`.
  contact:
    name: Institutional Kernel
  license:
    name: Uso interno — Mi Rotaract
    identifier: LicenseRef-MiRotaract-Internal
...
servers:
  - url: https://api.agendai.com.ar/api/kernel/v1
    description: Producción
  - url: https://staging-api.agendai.com.ar/api/kernel/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /persons:
    post:
      operationId: createPerson
      tags: [Persons]
      summary: Crear una persona (puede no tener cuenta)
      x-required-permission: kernel.person.manage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePersonRequest" }
      responses:
        "201":
          description: Creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
    get:
      operationId: listPersons
      tags: [Persons]
      summary: Listar personas
      x-required-permission: kernel.person.read
      parameters:
        - $ref: "#/components/parameters/query"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PersonPage" }
  /persons/{personId}:
    get:
      operationId: getPerson
      tags: [Persons]
      summary: Obtener una persona
      description: >
        Requiere `kernel.person.read`, o `kernel.person.read.self` si
        `personId` corresponde a la persona vinculada a la cuenta autenticada.
      x-required-permission: kernel.person.read
      parameters:
        - $ref: "#/components/parameters/personId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updatePerson
      tags: [Persons]
      summary: Actualizar datos de una persona
      description: >
        Requiere `kernel.person.manage`, o `kernel.person.update.self`
        sobre la propia persona.
      x-required-permission: kernel.person.manage
      parameters:
        - $ref: "#/components/parameters/personId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePersonRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
  /persons/{personId}/archive:
    post:
      operationId: archivePerson
      tags: [Persons]
      summary: Archivar una persona
      description: >
        Una persona archivada no puede recibir membresías, cargos ni roles
        nuevos (invariante 6.2.3). No elimina su historial (invariante 6.2.2).
      x-required-permission: kernel.person.manage
      parameters:
        - $ref: "#/components/parameters/personId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Person" }
  /persons/{personId}/invitations:
    post:
      operationId: invitePersonToCreateAccount
      tags: [Persons]
      summary: Invitar a una persona a vincular una cuenta
      x-required-permission: kernel.person.manage
      parameters:
        - $ref: "#/components/parameters/personId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [membershipId, email]
              properties:
                membershipId: { type: string }
                email: { type: string, format: email }
      responses:
        "201":
          description: Invitación creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountInvitation" }

  # ---------------------------------------------------------------------
  # 9.4 Organizaciones
  # ---------------------------------------------------------------------
...
  /organizations/{organizationId}/memberships:
    post:
      operationId: createMembership
      tags: [Memberships]
      summary: Crear una membresía en una organización
      x-required-permission: kernel.membership.create
      parameters:
        - $ref: "#/components/parameters/organizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMembershipRequest" }
      responses:
        "201":
          description: Creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "409":
          description: Ya existe membresía para esa persona en la organización (CA-MEM-02).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      operationId: listOrganizationMemberships
      tags: [Memberships]
      summary: Listar membresías de una organización
      x-required-permission: kernel.membership.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/MembershipStatus" }
          style: form
          explode: true
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipPage" }
  /memberships/{membershipId}:
    get:
      operationId: getMembership
      tags: [Memberships]
      summary: Obtener una membresía
      x-required-permission: kernel.membership.read
      parameters:
        - $ref: "#/components/parameters/membershipId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateMembership
      tags: [Memberships]
      summary: Actualizar metadatos de una membresía
      x-required-permission: kernel.membership.update
      parameters:
        - $ref: "#/components/parameters/membershipId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                memberNumber: { type: string, nullable: true }
                internalNotes: { type: string, nullable: true }
                metadata: { type: object, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
  /memberships/{membershipId}/activate:
    post:
      operationId: activateMembership
      tags: [Memberships]
      summary: Activar una membresía (PENDING → ACTIVE)
      description: Requiere `joinedAt` (invariante 6.4.2). Registra transición (CA-MEM-03).
      x-required-permission: kernel.membership.activate
      parameters:
        - $ref: "#/components/parameters/membershipId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                joinedAt: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /memberships/{membershipId}/leave:
    post:
      operationId: putMembershipOnLeave
      tags: [Memberships]
      summary: Pasar a licencia (ACTIVE → ON_LEAVE)
      x-required-permission: kernel.membership.update
      parameters:
        - $ref: "#/components/parameters/membershipId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reasonCode: { type: string, nullable: true }
                reasonText: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /memberships/{membershipId}/resume:
    post:
      operationId: resumeMembership
      tags: [Memberships]
      summary: Volver de licencia (ON_LEAVE → ACTIVE)
      x-required-permission: kernel.membership.update
      parameters:
        - $ref: "#/components/parameters/membershipId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /memberships/{membershipId}/deactivate:
    post:
      operationId: deactivateMembership
      tags: [Memberships]
      summary: Desactivar una membresía (→ INACTIVE)
      description: Requiere `endedAt` (invariante 6.4.3).
      x-required-permission: kernel.membership.deactivate
      parameters:
        - $ref: "#/components/parameters/membershipId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reasonCode: { type: string, nullable: true }
                reasonText: { type: string, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /memberships/{membershipId}/graduate:
    post:
      operationId: graduateMembership
      tags: [Memberships]
      summary: Marcar egreso (ACTIVE → GRADUATED)
      x-required-permission: kernel.membership.update
      parameters:
        - $ref: "#/components/parameters/membershipId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /memberships/{membershipId}/reactivate:
    post:
      operationId: reactivateMembership
      tags: [Memberships]
      summary: Reactivar una membresía en estado terminal (→ ACTIVE)
      description: >
        Preserva el mismo ID y agrega historial (invariante 6.4.7). Sólo
        aplica desde `INACTIVE` según la máquina de estados (§7.3); estados
        `GRADUATED`/`TRANSFERRED` no tienen retorno documentado y deberían
        resolverse con una nueva membresía o un `MembershipTransfer` (supuesto,
        pendiente de aclaración en `kernel-spec.md`).
      x-required-permission: kernel.membership.update
      parameters:
        - $ref: "#/components/parameters/membershipId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationMembership" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /memberships/{membershipId}/history:
    get:
      operationId: getMembershipHistory
      tags: [Memberships]
      summary: Historial de transiciones de una membresía
      x-required-permission: kernel.membership.read
      parameters:
        - $ref: "#/components/parameters/membershipId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MembershipTransition" }
  /persons/{personId}/memberships:
    get:
      operationId: listPersonMemberships
      tags: [Memberships]
      summary: Listar membresías de una persona
      description: >
        Requiere `kernel.membership.read`, o acceso propio si `personId`
        corresponde a la persona autenticada.
      x-required-permission: kernel.membership.read
      parameters:
        - $ref: "#/components/parameters/personId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrganizationMembership" }

  # ---------------------------------------------------------------------
  # 9.6 Períodos
  # ---------------------------------------------------------------------
  /organizations/{organizationId}/periods:
    post:
      operationId: createPeriod
      tags: [Periods]
      summary: Crear un período (DRAFT)
      x-required-permission: kernel.period.create
      parameters:
        - $ref: "#/components/parameters/organizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePeriodRequest" }
      responses:
        "201":
          description: Creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "422":
          description: "Fechas inválidas: el período debe ir del 1 de julio al 30 de junio siguiente (CA-PER-01/01a)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      operationId: listPeriods
      tags: [Periods]
      summary: Listar períodos de una organización
      x-required-permission: kernel.period.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/PeriodStatus" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InstitutionalPeriod" }
  /organizations/{organizationId}/periods/current:
    get:
      operationId: getCurrentPeriod
      tags: [Periods]
      summary: Obtener el período ACTIVE vigente
      x-required-permission: kernel.period.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "404":
          description: No hay período activo.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
  /periods/{periodId}:
    get:
      operationId: getPeriod
      tags: [Periods]
      summary: Obtener un período
      x-required-permission: kernel.period.read
      parameters:
        - $ref: "#/components/parameters/periodId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateDraftPeriod
      tags: [Periods]
      summary: Actualizar un período en estado DRAFT
      x-required-permission: kernel.period.update
      parameters:
        - $ref: "#/components/parameters/periodId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePeriodRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "409":
          description: El período no está en DRAFT.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Error" }
  /periods/{periodId}/schedule:
    post:
      operationId: schedulePeriod
      tags: [Periods]
      summary: Programar un período (DRAFT → SCHEDULED)
      x-required-permission: kernel.period.update
      parameters:
        - $ref: "#/components/parameters/periodId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /periods/{periodId}/activate:
    post:
      operationId: activatePeriod
      tags: [Periods]
      summary: Activar un período (SCHEDULED → ACTIVE)
      description: >
        Sólo se activa desde `SCHEDULED` (invariante 6.5.4). No puede haber
        dos períodos `ACTIVE` para la misma organización (invariante 6.5.2,
        CA-PER-02). Emite `kernel.period.activated.v1` (CA-PER-03).
      x-required-permission: kernel.period.activate
      parameters:
        - $ref: "#/components/parameters/periodId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /periods/{periodId}/close:
    post:
      operationId: closePeriod
      tags: [Periods]
      summary: Cerrar un período (ACTIVE → CLOSED)
      description: >
        Finaliza sus cargos activos en la misma transacción
        (invariante 6.5.7, CA-PER-04).
      x-required-permission: kernel.period.close
      parameters:
        - $ref: "#/components/parameters/periodId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "409": { $ref: "#/components/responses/InvalidTransition" }
  /periods/{periodId}/cancel:
    post:
      operationId: cancelPeriod
      tags: [Periods]
      summary: Cancelar un período (DRAFT/SCHEDULED → CANCELLED)
      description: >
        (Supuesto) `kernel-spec.md` §10.1 no define un permiso propio para
        cancelar; se asume que reutiliza `kernel.period.update` por ser una
        corrección administrativa previa a la activación.
      x-required-permission: kernel.period.update
      parameters:
        - $ref: "#/components/parameters/periodId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstitutionalPeriod" }
        "409": { $ref: "#/components/responses/InvalidTransition" }

  # ---------------------------------------------------------------------
  # 9.7 Cargos
  # ---------------------------------------------------------------------
  /position-definitions:
    post:
      operationId: createPositionDefinition
      tags: [Positions]
      summary: Registrar una definición de cargo
      description: >
        Registra un cargo. Todo cargo DISTRICT configurable debe indicar el
        distrito propietario y un permiso de edición. La autorización se
        evalúa en ese distrito.
      x-required-permission: kernel.position.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/CreatePositionDefinitionRequest" }
      responses:
        "201":
          description: Creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionDefinition" }
    get:
      operationId: listPositionDefinitions
      tags: [Positions]
      summary: Listar definiciones de cargo
      x-required-permission: kernel.position.read
      parameters:
        - name: organizationType
          in: query
          schema: { $ref: "#/components/schemas/OrganizationType" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PositionDefinition" }
  /position-definitions/{positionDefinitionId}:
    patch:
      operationId: updatePositionDefinition
      tags: [Positions]
      summary: Editar una definición de cargo
      description: >
        Requiere el editPermissionCode configurado en el cargo y se evalúa
        sobre el distrito propietario. Los cargos de sistema no son editables
        desde un distrito.
      x-required-permission: dynamic:position.editPermissionCode
      parameters:
        - $ref: "#/components/parameters/positionDefinitionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/UpdatePositionDefinitionRequest" }
      responses:
        "200":
          description: Actualizada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PositionDefinition" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /position-definitions/{positionDefinitionId}/permissions/{permissionId}:
    put:
      operationId: attachPermissionToPosition
      tags: [Positions]
      summary: Asociar un permiso al rol técnico de un cargo
      description: >
        Requiere editPermissionCode. Actualiza el rol defaultRoleCode del
        cargo; responde 409 si el cargo no tiene rol técnico asociado.
      x-required-permission: dynamic:position.editPermissionCode
      parameters:
        - $ref: "#/components/parameters/positionDefinitionId"
        - $ref: "#/components/parameters/permissionId"
      responses:
        "204": { description: Permiso asociado. }
        "409": { $ref: "#/components/responses/InvalidTransition" }
    delete:
      operationId: detachPermissionFromPosition
      tags: [Positions]
      summary: Quitar un permiso del rol técnico de un cargo
      x-required-permission: dynamic:position.editPermissionCode
      parameters:
        - $ref: "#/components/parameters/positionDefinitionId"
        - $ref: "#/components/parameters/permissionId"
      responses:
        "204": { description: Permiso quitado. }
        "409": { $ref: "#/components/responses/InvalidTransition" }
...
  /organizations/{organizationId}/authorities/current:
    get:
      operationId: getCurrentAuthorities
      tags: [Appointments]
      summary: Obtener las autoridades ACTIVE del período vigente
      description: >
        `Appointment` es la única fuente de verdad para presidencias y
        autoridades (§4.5, CA-APP-03, CA-APP-04). No existen flags
        `isPresident`.
      x-required-permission: kernel.appointment.read
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [Appointments]
      summary: Obtener un cargo
      x-required-permission: kernel.appointment.read
      parameters:
        - $ref: "#/components/parameters/appointmentId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token de usuario emitido por `POST /auth/login` o
        `/auth/refresh` (§14.2). Vida útil de 10 minutos. `aud`:
        `mirotaract-platform`.
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token de servicio a servicio firmado (o mTLS equivalente, §14.3).
        `sub` identifica al servicio emisor. `aud` debe ser
        `institutional-kernel`. Requerido para todos los endpoints bajo
        `/service/*`; el kernel rechaza estas rutas con token de usuario.
  parameters:
...
    query:
      name: query
      in: query
      schema: { type: string }
      description: Búsqueda de texto libre.
...